requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
webdriver-manager==4.0.1
python2captcha==1.1.3
PyPDF2==3.0.1
Pillow==10.1.0
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0